Add Adhiti's initial content draft again (#2)
This is a duplicate of 5f6ab30 plus an extra commit in 51ae313 based on @shwina review of PR #2.

* Add blank line to README.md

* Revert "Add blank line to README.md"

This reverts commit 7a306afd71acbc1de76b710c1c314a5e260404e5.

* Add a template for episode

This file simply collects some useful markdown formatting markup in one place. This also captures the title block structure that's included in the beginning of each episode file.

* Add Draft1_Adhiti.md content to _episodes

I took the liberty to put Adhiti's proposed structure into the episode format. Not sure if this actually works yet. We'll see once I merge this into gh-pages on my fork.

* Update 14-ros-multiple-machines.md

* Create sources_list.md

* fix missing inline code markup

Thanks to @shwina review of PR #2